updated userchange event.
I do not know if this fixes the issue that was occurring last night where !roll was inappropriately moving medics into incorrect server rooms, but this definitely fixes an error where channel lengths didn't subtract if you moved into a 'sibling' channel in a server room. Theoretically that will fix it.
Also added pcl() function that I used to fix the problem, but commented out. Uncomment line 367 for verbose logging on channel lengths.